Fix for fullscreen in HiDPI monitors: use scaled resolution, not highest
By default, FULLSCREEN mode would use the highest resolution supported
by the monitor. For example, 5120x2880 on 5K displays, however,
according to [1], the NSHighResolutionCapable key is needed to do this.

Since HiDPI is overkill, and complicates application packaging
requirements, instead use the scaled-down resolution. Finding this is
somewhat complicated, requiring creating a window and comparing vs
glfwGetFramebufferSize(), detailed in [2], but getting the scaling
factor is already implemented in Craft in get_scale_factor(). This
function requires a window to get the scale factor of, so create it in
g->window, get the scale factor, then destroy it and recreate with the
smaller window size.

[1] http://www.glfw.org/docs/latest/group__window.html#ga5c336fddf2cbb5b92f65f10fb6043344
[2] http://stackoverflow.com/questions/25230841/how-to-find-display-scaling-factor-on-retina-4k-displays