vulkanize: Convert shaders to explicit locations
Spir-v requires explicit locations. Additionally, this allows avoiding
the location and binding queries for shader attributes and uniforms that
were used before. To use the explicit binding feature, version 420 of
the shading language is required.

Also rename attribute and varying keywords to the more modern in/out
qualifiers and eliminate the builtin color output variable.

As a side effect, the attrib object is not needed for many drawing
functions. So it's removed from them.