vulkanize: Abstract buffer object creation
To allow the same controlling code in main to perform the same rendering
functions by passing around buffer objects, this abstracts the
representation of the buffer object away as a pointer to a partially
defined struct. In the case of GL, this just contains an integer
identifier. For Vulkan, it will carry a lot more.

[Resolved merge conflicts during cherry picking.]