vulkanize: Precreate vertex buffers for reuse
Most buffers are created and destroyed immediately before and after they
are used for rendering. This is functional, though inefficient for
OpenGL, but it causes more serious problems in vulkan due to its greater
asynchronicity.

To accomplish this, the wireframe vertex buffer lost its position
information in favor of adding it to the model matrix at render time.
For crosshairs, the dimensions of the frame are extracted a little
earlier and used to determine its location.

Instead of allocating buffers for item rendering on demand, since the
location and geometry doesn't change, the buffers can be precreated and
reused as many times as needed