vulkanize: Add comments to renderer code
I have attempted to maintain the implicit style included in this
project. However, I don't feel that a complete lack of comments is a
style. Since I intend to document my later contributions to the renderer
with extensive comments, I think it would be more consistent if the
previous code was similarly documented.

I'm not going to document the entire project, but this adds guidance for
the portions relevant to rendering