Generate trees of different sizes
Parameterizes the tree size and use deterministic simplex calls to
determine for any given tree placement which size of tree to attempt to
place. The parameterization is a little phenomenological as opposed to
from first principles, but oh well.

In practical terms, where the old code rigidly generated trees of size
"4" (in terms of the chosen parameter), this code now generates trees of
sizes 3 to 5 (driving both length of stem and size of crown
simultaneously). Larger trees look way too regular and artificial,
smaller than 3 no longer works in terms of the algorithm (no crown
left), but also wouldn't actually make sense to try and model the crown
as a ~half sphere anyway.

This also moves the tree generation into its own single purpose
file-local function.