Defensiveness against negative w
The exact purpose of sometimes flipping the block type parameter "w" to
a negative value of the same absolute value is still a touch unclear to
me. But it seems quite clear that it's being used to transmit an
orthogonal (literal) bit of information from one place to some other
without creating separate parameter passing hassle. Meh. The problem is
that I haven't found yet where that's actually used. Unsure how to grep
for that so need to read more code.

Nonetheless, that information is quite surely NOT used in the predicate
functions that operate on item types, thus theys hould all have an abs()
call at the start. I suspect this is not an actively impacting bug but
rather a latent one, though. I wonder if I'm going to be reverting this
commit later. :)