Introduce a typedef for block types (ya know, the ole "int w")
This is almost certainly not complete, and it's also a functional change
in that types were stored as char's in the map structs to save memory.
Should probably measure the change in memory consumption, but tbh the
"char w" lived right next to 3 "int x,z,y" and padding probably meant
that it was using all the memory of a full native int size anyway.