Fix issue in window resizing
The crosshair buffer was created statically to the window size. So
changing to fullscreen (and back) would mess it up. This keeps track of
the window size last time the crosshair buffer was generated and if they
differ, regenerates the buffer.

This is ... ugly... but given my lack of opengl (nor vulkan) knowledge
and a weird behavior where window sizes returned by library calls appear
to update with delay, this seems an acceptable workaroung. I suppose one
could make the buffer "dynamic" and update it instead, but given that
window size changes have to be considered really rare, that's hardly
worth it.